Main activity and main screen changes
Main activity now has automatic check to see if a user is logged in or out, if the user is logged in, he goes to main screen, if the user is logged out he goes to the log in screen.

Added a button to log out in main screen, will be moved to a different section but the implementation of it is there.