NFC implementation
this implements the NFC functionality to the application, code is very complex but NFC is now enabled in the AndroidManifest.xml, created another XML called apduservice, this one defines the AID and their intents. HostCardEmulationService.kt is in charge of replying and validating. Utils.kt has been copied from online but it basically translates hex strings to bytes in order to validate data easier (helped methods).